+++ atompack/errors.py
@@ -1,0 +1,2 @@
+class CrystallographyError(Exception):
+    pass